Try to reduce issues where lights get stuck on dim
This happens when making a change to `light.bedroom` which then goes on
to adjust the brightness for several other lights in the bedroom we are
listening to. We receive events for them but Home Assistant doesn't
realise it was triggered by us. So the dimming cancels the lights being
turned off.